+++ tests/test_ip.py
@@ -1,4 +1,8 @@
 
-def test_rdap_asn_lookup(rdapc):
-    obj = rdapc.get_ip("1.1.1.1")
+def test_rdap_ip_lookup(rdapc):
+    query = "1.1.1.1"
+    obj = rdapc.get_ip(query)
     assert obj.data
+
+    # test get duduction
+    assert obj.data == rdapc.get(query).data